+++ app/src/main/java/com/github/forrestdp/bankingapp/network/bankinginfo/Transaction.kt
@@ -2,6 +2,7 @@ package com.github.forrestdp.bankingapp.network.bankinginfo
 
 import com.github.forrestdp.bankingapp.utils.CurrencyCode
 import com.squareup.moshi.Json
+import java.math.BigDecimal
 import java.util.*
 
 data class Transaction(
@@ -10,7 +11,7 @@ data class Transaction(
     @Transient
     val currencyCode: CurrencyCode = CurrencyCode.USD,
     @Transient
-    val currencyMultiplier: Double = 1.0,
+    val currencyMultiplier: BigDecimal = 1.0.toBigDecimal(),
     val title: String,
     @Json(name = "icon_url")
     val iconUrl: String,
