+++ app/src/main/java/com/github/forrestdp/bankingapp/network/currencyinfo/Currency.kt
@@ -1,12 +1,13 @@
 package com.github.forrestdp.bankingapp.network.currencyinfo
 
 import com.squareup.moshi.Json
+import java.math.BigDecimal
 
 data class Currency(
     @Json(name = "Value")
-    val value: Double
+    val value: BigDecimal
 ) {
     companion object {
-        val defaultCurrency = Currency(1.0)
+        val defaultCurrency = Currency(1.0.toBigDecimal())
     }
 }
